requirements.txt
aiogram==3.0.0b7
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.2.0
APScheduler==3.9.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
environs==9.5.0
frozenlist==1.3.1
idna==3.4
marshmallow==3.18.0
multidict==6.0.2
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
redis==4.5.3
requests==2.28.1
six==1.16.0
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
yarl==1.8.1

betterlogging~=0.2.1
alembic~=1.11.1
asyncpg~=0.27.0
pydantic~=1.10.7
SQLAlchemy~=2.0.15
google-api-python-client>=2.86.0